Add functions for playing rock paper scissors
This commit adds functions for playing rock paper scissors. It includes
a function to randomly generate a computer choice, a function to prompt
the user for their choice, variables for tracking score, and a function
for playing a single round (I probably should have committed them one
at a time, but I locked in too hard - oops). This commit is equivalent
to steps 1-5 of the Rock Paper Scissors project from The Odin Project.