Add psuedocode for function playGame
This commit adds psuedocode for function playGame, which will run
playRound 5 times and then compare the human score with the computer
score and declare a winner. I'm running out of time in my day, and I
don't think I'll be able to finish this until tomorrow, so I'm making
sure I commit this evening just to be safe (especially since I forgot
to commit anything until after step 5 of the project). I'll be able to
finish this tomorrow on my own time for sure - I think step 5 was the
hard part; barring some severe underestimation on my part, this last
step should be relatively easy.