Add function playGame that plays 5 rounds
This commit adds the function playGame, which runs the playRound
function 5 times and then compares score. It also makes some small
adjustments to other functions in order to work properly with the new
function. Naturally, I have come to regret my hubris from the last
commit, in which I claimed that this would not take long. In particular
I had some difficulty with reading and writing variables to and from
different scopes - I didn't realize during the lesson how much trouble
this would end up causing me. But now it works, and it's done.