Add ui and refactor scripts to work with html
This commit adds an HTML-based UI, and refactors some of the scripts to
accommodate. The user now selects their choice via button, and clicking
the button now passes a value to the humanVar and runs the playRound
function with humanVar as a parameter. The playRound function logs the
result in a list item element, and the humanScore and computerScore
variables are display in a paragraph element. Once either humanScore or
computerScore reaches a value of 5, the playRound function will display
a game over message that says who won, then reset the scores. This
renders the playGame function redundant, and thus it was removed.
Finally, I added a button that clears all the log messages from the
page.